Migration to ts - 기존에 js로 선언된 source file에 ts 적용
