--- src/api/index.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const user = require('./controller/user.ctrl');
-
-router.use('/user', user);
-
-module.exports = router;
