+++ src/api/index.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import user from './controller/user.ctrl';
+
+const router = express.Router();
+
+router.use('/user', user);
+
+export default router;
