+++ src/config/index.ts
@@ -1,17 +1,16 @@
-const dotenv = require('dotenv');
-const path = require('path');
+import dotenv from 'dotenv';
+import path from 'path';
 
 dotenv.config({ path: path.resolve(__dirname, '.env') });
 
-const config = {
-    development: {
-        port: process.env.PORT,
-        dialect: process.env.DB_DIALECT,
+export default {
+    port: process.env.PORT,
+    db: {
+        type: process.env.DB_TYPE,
         host: process.env.DB_HOST,
+        port: process.env.DB_PORT,
         database: process.env.DATABASE,
         username: process.env.DB_USERNAME,
         password: process.env.DB_PASSWORD,
     },
 };
-
-module.exports = config;
