+++ src/lib/resResult.ts
@@ -1,0 +1,8 @@
+export default (status = false, code = 400, message = '', data: any = null) => {
+    return {
+        status,
+        code,
+        message,
+        data,
+    };
+};
