+++ src/socket.ts
@@ -2,17 +2,17 @@ const app = require('express')();
 const server = require('http').createServer(app);
 const io = require('socket.io')(server);
 
-app.set('view engine', 'ejs'); 
-app.set('views',  __dirname + '/views');
+app.set('view engine', 'ejs');
+app.set('views', __dirname + '/views');
 
 app.get('/', (req, res) => {
     res.render('test.ejs');
-})
+});
 
 io.on('connection', (socket) => {
     //socket.emit으로 현재 연결한 상대에게 신호를 보낼 수 있다.
-    socket.emit('usercount', io.engine.clientsCount);    
-    
+    socket.emit('usercount', io.engine.clientsCount);
+
     socket.on('usercount', (count) => {
         io.emit('count', count);
     });
